Refactor API code to new package. Create 'game' events.
