+++ src/engine/creature-db.ts
@@ -5,6 +5,7 @@ import { filter, findIndex, flow, keys, map, reduce } from 'lodash/fp'
 import { dartLizard, DefaultDartLizardSpeed } from '../game-content/scripts/creatures/dart-lizard'
 import { player } from '../game-content/scripts/creatures/player'
 
+import { CreatureScript } from './api/creature-script'
 import { attackPlayer } from './behaviors/attack'
 import { BehaviorChain } from './behaviors/behavior-chain'
 import { guard } from './behaviors/guard'
@@ -18,7 +19,6 @@ import { MonsterLootTables } from './data/loot-tables'
 import { Item } from './item'
 import { ItemTemplate } from './item-db'
 import { getTraversableNeighbors } from './map/map-utils'
-import { CreatureScript } from './script-api'
 import { ageSensor } from './sensors/age-sensor'
 import { allCreaturesSensor, creaturesInFrontSensor, getDetectedCreatures } from './sensors/creature-sensors'
 import { facingSensor } from './sensors/facing-sensor'
