+++ src/engine/events/creature.ts
