+++ src/engine/ui-controller.ts
@@ -1,0 +1,4 @@
+export interface UiController {
+  /** indicates if the UI is ready (visible, responding to events, etc.) */
+  readonly ready: boolean
+}
