+++ src/engine/world.ts
@@ -6,7 +6,7 @@ import { getResultMessage } from './actions/result-handler'
 import { Creature } from './creature'
 import { Dungeon } from './dungeon/dungeon'
 import { WorldEvents } from './events'
-import { CreatureEvents } from './events/creature-events'
+import { CreatureEvents } from './events/creature'
 import { Item } from './item'
 import { ExpeditionMap } from './map/map'
 import { Player } from './player'
