+++ src/game-content/scripts/creatures/dart-lizard.ts
@@ -1,7 +1,9 @@
+import { CreatureScript } from 'engine/api/creature-script'
+import { ScriptApi } from 'engine/api/script-api'
+import { WorldScript } from 'engine/api/world-script'
 import { Item } from 'engine/item'
 import { getAdjacentCoordinates } from 'engine/map/map-utils'
 import { random } from 'engine/random'
-import { CreatureScript, ScriptApi, WorldScript } from 'engine/script-api'
 import { getFacing } from 'engine/sensors/facing-sensor'
 import { getStartledTurns, isStartled } from 'engine/sensors/startle-sensor'
 import { countBy, get, some } from 'lodash/fp'
